Open multiqueue tun device even with 1 queue
Make the use of the IFF_MULTI_QUEUE flag independent of the number of
queues actually requested.

It is possible and valid to open a tun device in multiqueue mode but
still open only one queue e.g. if the tunnel device was created for you
as multiqueue but you only want to access it on a single thread, or if
you want to open more queues later.